// src/components/burger-ingredient/burger-ingredient.types.ts
export type Ingredient = {
  _id: string;
  name: string;
  type: string;
  proteins: number;
  fat: number;
  carbohydrates: number;
  calories: number;
  price: number;
  image: string;
  image_mobile: string;
  image_large: string;
  __v: number;
};

export type BurgerIngredientComponent = {
  ingredient: Ingredient;
  className?: string;
};